cambio de base de datos mysql a oracle

+++ src/main/java/com/demo/demo/Features/Param/User/Controllers/UserController.java
@@ -6,9 +6,13 @@ import com.demo.demo.Features.Param.User.Services.UserService;
 import com.demo.demo.Features.Param.User.Validations.UserValidations;
 import com.demo.demo.models.*;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,18 +27,22 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    Logger logger = LoggerFactory.getLogger(UserService.class);
+
+
     @GetMapping
     @Tag(name = "Usuarios", description = "Controlador de usuarios")
     @Operation(summary = "Obtender todos los usuarios",
             responses = {
             @ApiResponse(responseCode = "200", description = "usuarios obtenidos exitosamente"),
-            @ApiResponse(responseCode = "400", description = "no hay usuarios para mostrar")
+            @ApiResponse(responseCode = "400", description = "no hay usuarios para mostrar", content = @Content(schema = @Schema(implementation = NoDataResponse.class)))
         }
     )
     public ResponseEntity<Response<ArrayList<GetUserResponseAll>>> getUsers(
         @RequestParam(value = "nombre", required = false) String nombre,
         @RequestParam(value = "apellido", required = false) String apellido
     ){
+        logger.info("Consultando datos de usuarios");
 
         ArrayList<GetUserResponseAll> result = this.userService.getUsers(nombre,apellido);
         Response<ArrayList<GetUserResponseAll>> dataResponse = new Response<>();
@@ -58,7 +66,7 @@ public class UserController {
         summary = "Obtender un usuario por su id",
         responses = {
             @ApiResponse(responseCode = "200", description = "usuario obtenido correctamente"),
-            @ApiResponse(responseCode = "400", description = "el usuario no existe")
+            @ApiResponse(responseCode = "400", description = "el usuario no existe", content = @Content(schema = @Schema(implementation = NoDataResponse.class)))
         }
     )
     public ResponseEntity<Response<GetUserResponseById>> getUserById(@PathVariable Long id){
@@ -87,16 +95,6 @@ public class UserController {
     public ResponseEntity<Response<SaveUserResponse>> saveUser(@Valid @RequestBody SaveUserRequest user, BindingResult bindingResult){
         Response<SaveUserResponse> dataResponse = new Response<>();
 
-        UserModel userModel = new UserModel();
-        userModel.setFirsName(user.getFirsName());
-        userModel.setLastName(user.getLastName());
-        userModel.setEmail(user.getEmail());
-
-        CityModel cityModel = new CityModel();
-        cityModel.setId(user.getCiudad());
-
-        userModel.setCityId(cityModel);
-
         List<String> Errors = UserValidations.handleValidationErrors(bindingResult);
 
         if(!Errors.isEmpty()){
@@ -108,7 +106,7 @@ public class UserController {
 
         try {
 
-            SaveUserResponse result = this.userService.saveUser(userModel);
+            SaveUserResponse result = this.userService.saveUser(user);
 
             dataResponse.setStatus("Ok");
             dataResponse.setData(result);
@@ -133,7 +131,6 @@ public class UserController {
         Response<UpdateUserResponse> dataResponse = new Response<>();
 
         List<String> Errors = UserValidations.handleValidationErrors(bindingResult);
-        System.out.println("HOLA");
 
         if(!Errors.isEmpty()){
             dataResponse.setStatus("Error");
@@ -162,13 +159,32 @@ public class UserController {
     @Tag(name = "Usuarios", description = "Controlador de usuarios")
     @Operation(summary = "Eliminar usuario",
             responses = {@ApiResponse(responseCode = "200", description = "Operación exitosa")})
-    public String deleteUserById (@PathVariable Long id){
+    public ResponseEntity<Response<String>> deleteUserById (@PathVariable Long id){
+
+        Response<String> dataResponse = new Response<>();
+
+
+        GetUserResponseById response  = this.userService.getUserById(id);
+
+        if(response == null){
+            dataResponse.setStatus("Error");
+            dataResponse.setData(null);
+            dataResponse.setMessage("El usuario " + id + " no existe");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(dataResponse);
+        }
+
         boolean ok = this.userService.deleteUserById(id);
 
         if(ok){
-            return "Usuario" + id +" eliminado exitosamente";
+            dataResponse.setStatus("Ok");
+            dataResponse.setData(null);
+            dataResponse.setMessage("Usuario: " + id + " eliminado exitosamente");
+            return ResponseEntity.ok(dataResponse);
         }else{
-            return "No de puso eliminar el usuario" + id;
+            dataResponse.setStatus("Error");
+            dataResponse.setData(null);
+            dataResponse.setMessage("No fue posible eliminar el usuario");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(dataResponse);
         }
     }
 }
